requirements.txt
numpy
sounddevice
librosa
fastdtw
torch
noisereduce
colorama
paho-mqtt
langchain
redis
chromadb
ollama
ascii-magic
websockets
dtw
SpeechRecognition
subprocess
threading
noisereduce
langchain-community
scipy
pyaudio
scikit-learn
